# src/main/resources/openapi/api.yaml
openapi: 3.0.3

info:
  title: Multi Persistence Backend
  version: 1.0.0
  description: Backend demo for comparing relational and NoSQL databases

servers:
  - url: http://localhost:8080

tags:
  - name: Users
    description: User management APIs
  - name: Products
    description: Product management APIs
  - name: Orders
    description: Order management APIs
  - name: Order Items
    description: APIs for managing order items
  - name: Stats
    description: Aggregation and reporting APIs

paths:

  /users:
    get:
      tags: [Users]
      summary: Get all users
      description: Returns the list of all users.
      operationId: getUsers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'

    post:
      tags: [Users]
      summary: Create user
      description: Creates a new user.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user by id
      description: Returns the details of a user by id.
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      summary: Delete user by id
      description: Deletes a user by id.
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User deleted

  /products:
    get:
      tags: [Products]
      summary: Get all products
      description: Returns the list of all products.
      operationId: getProducts
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'

    post:
      tags: [Products]
      summary: Create product
      description: Creates a new product.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /products/{productId}:
    get:
      tags: [Products]
      summary: Get product by id
      description: Returns the details of a product by id.
      operationId: getProductById
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Products]
      summary: Update product
      description: Updates the name and price of an existing product.
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags: [Products]
      summary: Delete product by id
      description: Deletes a product by id.
      operationId: deleteProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: Product deleted

  /orders:
    get:
      tags: [Orders]
      summary: Get orders
      description: Returns all orders.
      operationId: getOrders
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'

    post:
      tags: [Orders]
      summary: Create order
      description: Creates a new order for a user.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/InternalServerError'

  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by id
      description: Returns the details of an order by id.
      operationId: getOrderById
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Orders]
      summary: Delete order by id
      description: Deletes an order by id.
      operationId: deleteOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '204':
          description: Order deleted

  /orders/user/{userId}:
    get:
      tags: [Orders]
      summary: Get orders by user id
      description: Returns all orders created by the specified user.
      operationId: getOrdersByUserId
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: List of orders for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /orders/stats/total-spent-per-user:
    get:
      tags: [Stats]
      summary: Get total spent per user
      description: Returns the total amount spent by each user.
      operationId: getTotalSpentPerUser
      responses:
        '200':
          description: Total spent per user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TotalSpentPerUserResponse'

  /orders/stats/most-sold-products:
    get:
      tags: [Stats]
      summary: Get most sold products
      description: Returns products sorted by sold quantity.
      operationId: getMostSoldProducts
      responses:
        '200':
          description: Most sold products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MostSoldProductResponse'

components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      description: User identifier
      schema:
        type: string
        format: uuid

    ProductId:
      name: productId
      in: path
      required: true
      description: Product identifier
      schema:
        type: string
        format: uuid

    OrderId:
      name: orderId
      in: path
      required: true
      description: Order identifier
      schema:
        type: string
        format: uuid

    ItemId:
      name: itemId
      in: path
      required: true
      description: Order item identifier
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:

    ErrorResponse:
      type: object
      required:
        - status
        - error
        - message
        - path
        - timestamp
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 404
        error:
          type: string
          description: Error type
          example: Not Found
        message:
          type: string
          description: Detailed error message
          example: User not found with id 123e4567-e89b-12d3-a456-426614174000
        path:
          type: string
          description: Request path
          example: /users/123e4567-e89b-12d3-a456-426614174000
        timestamp:
          type: string
          format: date-time
          description: Error timestamp

    UserRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        firstName:
          type: string
          description: User first name
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          description: User last name
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          description: User email
          minLength: 5
          maxLength: 255

    ProductRequest:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          description: Product name
          minLength: 1
          maxLength: 150
        price:
          type: number
          format: double
          description: Current product price
          minimum: 0

    OrderRequest:
      type: object
      required:
        - userId
        - items
      properties:
        userId:
          type: string
          format: uuid
          description: User identifier
        items:
          type: array
          description: List of items included in the order
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          format: uuid
          description: Product identifier
        quantity:
          type: integer
          description: Quantity of the product
          minimum: 1

    OrderItemUpdateRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
          description: Updated quantity
          minimum: 1

    UserResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
      properties:
        id:
          type: string
          format: uuid
          description: User identifier
        firstName:
          type: string
          description: User first name
        lastName:
          type: string
          description: User last name
        email:
          type: string
          format: email
          description: User email

    ProductResponse:
      type: object
      required:
        - id
        - name
        - price
      properties:
        id:
          type: string
          format: uuid
          description: Product identifier
        name:
          type: string
          description: Product name
        price:
          type: number
          format: double
          description: Current product price

    OrderResponse:
      type: object
      required:
        - id
        - user
        - items
        - total
      properties:
        id:
          type: string
          format: uuid
          description: Order identifier
        user:
          allOf:
            - $ref: '#/components/schemas/OrderUserResponse'
          description: User who created the order
        items:
          type: array
          description: List of order items
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        total:
          type: number
          format: double
          description: Total order amount
          minimum: 0

    OrderUserResponse:
      type: object
      required:
        - userId
        - firstName
        - lastName
        - email
      properties:
        userId:
          type: string
          format: uuid
          description: User identifier
        firstName:
          type: string
          description: User first name
        lastName:
          type: string
          description: User last name
        email:
          type: string
          format: email
          description: User email

    OrderItemResponse:
      type: object
      required:
        - productId
        - name
        - quantity
        - price
      properties:
        productId:
          type: string
          format: uuid
          description: Product identifier
        name:
          type: string
          description: Product name
        quantity:
          type: integer
          description: Quantity of the product
          minimum: 1
        price:
          type: number
          format: double
          description: Product price snapshot at order time
          minimum: 0

    TotalSpentPerUserResponse:
      type: object
      required:
        - userId
        - firstName
        - lastName
        - email
        - totalSpent
      properties:
        userId:
          type: string
          format: uuid
          description: User identifier
        firstName:
          type: string
          description: User first name
        lastName:
          type: string
          description: User last name
        email:
          type: string
          format: email
          description: User email
        totalSpent:
          type: number
          format: double
          description: Total amount spent by the user
          minimum: 0

    MostSoldProductResponse:
      type: object
      required:
        - productId
        - name
        - totalQuantity
      properties:
        productId:
          type: string
          format: uuid
          description: Product identifier
        name:
          type: string
          description: Product name
        totalQuantity:
          type: integer
          format: int64
          description: Total sold quantity
          minimum: 0